Move edit history page to modules
