+++ src/apps/investments/constants.js
@@ -170,14 +170,6 @@ const QUERY_DATE_FIELDS = [
   'likelihood_to_land',
 ]
 
-const CHANGE_TYPE_TEXT = 'the project'
-const TRUE = 'True'
-const FALSE = 'False'
-const NOT_SET = 'Not set'
-
-// TODO: When receive list of updated edit-history fields from users,
-// add EXCLUDED_FIELDS, FIELD_NAME_TO_LABEL_MAP
-
 module.exports = {
   GLOBAL_NAV_ITEM,
   LOCAL_NAV,
@@ -185,10 +177,6 @@ module.exports = {
   APP_PERMISSIONS,
   QUERY_FIELDS,
   QUERY_DATE_FIELDS,
-  CHANGE_TYPE_TEXT,
-  TRUE,
-  FALSE,
-  NOT_SET,
   LARGE_INVESTMENT_PROFILE_QUERY_FIELDS,
   LARGE_INVESTMENT_PROFILE_QUERY_DATE,
   LARGE_INVESTMENT_OPPORTUNITY_QUERY_FIELDS,
