+++ src/client/routes.js
@@ -40,7 +40,7 @@ import ProjectPropositions from './modules/Investments/Projects/ProjectPropositi
 import EditClientRelationshipManagement from './modules/Investments/Projects/EditClientRelationshipManagement'
 import EditProjectManagement from './modules/Investments/Projects/EditProjectManagement'
 import EditTeamMembers from './modules/Investments/Projects/EditTeamMembers'
-import InvestmentEditHistory from '../apps/investments/client/InvestmentEditHistory'
+import ProjectEditHistory from './modules/Investments/Projects/ProjectEditHistory'
 
 const routes = {
   companies: [
@@ -291,7 +291,7 @@ const routes = {
     {
       path: '/investments/projects/:projectId/edit-history',
       module: 'datahub:investments',
-      component: InvestmentEditHistory,
+      component: ProjectEditHistory,
     },
   ],
 }
